Make it possible to publish opensource repos in subdirectories (#19)
Change the github action we use to publish repos from the opensource
monorepo, to support publishing repos in subdirectories.

With this change we'll be able to publish `typeid/typeid-go` to a
`typeid-go` repo. It also makes it possible to specify a custom name
that does *not* match the directory name for cases when that's needed
(example `dir1/dir2:reponame`).

A follow up PR will use this github action to publish the typeid code.