Implement auth library using OIDC (#110)
## Summary
First pass at implementing a standalone authentication library for
command line tools. The idea is to open source it, in addition to using
it in our own tools. The library implements the authentication flow with
PCKE using OIDC.

At the moment this just has the basic flow, but I still need to add:
- Storage or caching of tokens
- Spinning up a server locally to listen for the flow's callback