[envsec] Improve performance (#129)
## Summary

This improves performance by making 2 changes:

* Params are looked up by path instead of by tags. Backwards
compatibitlity is perseved by doing tag looking for clients that specify
custom paths (e.g. launchpad)
* Cache aws credentials. When the cognito/STS exchange happens we get
credentials that are good for 60 minutes. We cache them locally and
reuse them if they exist and are not expired.

Total improvement was about 1.7s when credentials are cached (2s ->
300ms)

Other changes:

* Added xdg lib as internal (we can opensource so we can share among
projects)
* Added simple filecache lib (launchpad uses a similar version. We could
open source this as well)

## How was it tested?

* Used env variables to set auth0 as auth provider
* Benchmarked

When credentials are cached:

* envsec ls is ~500ms
* envsec ls for a single env is 200-300ms

Delta:
* Fetching credentials takes ~1.5s
* Switching from listByTags to listByPath saves 100-200ms