[easy][important] Fix build by propagating failures per go module (#190)
## Summary
Our current devbox scripts use find to run commands inside each of our
go modules. However, as written, the scripts return
the exit code of find and not of the executed command. So, if tests fail
in one of the go modules, the devbox script still succeeds, because it
"found" what we asked it to find.

Rewrite the scripts using `xargs` so that the exit code is correctly
propagated.

## How was it tested?
Broke a test, verified it propagates.