Flatten /pkg/sandbox to just /pkg (#192)
## Summary
Move all packages inside of `/pkg/sandbox/` to just `/pkg/`. The idea is
to simplify things:
+ Start by creating packages in `/pkg/`
+ If we decide to publish them on their own, move them to the top level.

In the process I renamed the `typeids` package to just `id`. I was
finding the collision with `typeid` too confusing, and I think being
able to do `id.OrgID` in the code is nicer.

## How was it tested?
`devbox run test`