[typeid-js] Enforce prefix (optionally) (#344)
## Summary

Brings it inline with unboxed version. Makes type-checking much more
reliable (because it now throws an error if prefix is wrong). Because
the current implementation uses `as T`, if the prefix is not specified,
the typechecker type has no relationship to the input data.

Arguably this should have been the default. It requires the same amount
of typing because the template type doesn't need to be added:

```ts
const id = TypeID.fromString(stringId, "foo")
// id is of type TypeID<"foo">
```

Unfortunately we can't do this without breaking backward compatibility. 

## How was it tested?

unit tests