[runx] Fix platform compatibility check for artifacts (#401)
## Summary
This fixes two errors in the `findArtifactForPlatform` function:
* If the artifact from the GitHub repository is not compatible with the invoker's platform, now it correctly returns a `types.ErrPlatformNotSupported` error, and it returns a `types.NoKnownArchive` error in case the artifact isn't packaged in a known artifact format
* The way the artifact name string was parsed wasn't working with `x86_64` architecture: the string was split in two because of the underscore, thus never matching correctly. I simplified the logic, by simply checking if the artifact name string, forced to lowercase, contains the invoker's OS and architecture

## How was it tested?
The first issue was tested by creating an ad-hoc GitHub project, compiling it for windows only, and uploading the compiled artifact to a GitHub release, and trying to donwload the binary via both standalone `runx` and with a local `devbox` build from a `linux/amd64` machine. The `NoKnownArchive` error was tested by trying to download an ad-hoc private `.deb` package.

The second issue was tested by downloading the package `pb33f/wiretap@latest` from GitHub, previously not possible, with both standalone `runx` and a local `devbox` build.