[runx] Allow unknown file types (#403)
## Summary

This tweaks the logic in
https://github.com/jetify-com/opensource/pull/401 to allow unknown file
types if there are no known file types in the list of released
artifacts. This is needed because some releases don't have file types at
all (e.g. https://github.com/mvdan/gofumpt/releases). We try to return a
known archive, otherwise we just return the last artifact that matches
the platform (if any)

Note: this is a revert to how the logic behaved before
https://github.com/jetify-com/opensource/pull/401 but keeps the artifact
name improvements from that PR.

cc: @edorizzardi

## How was it tested?

* `runx --install mvdan/gofumpt`
* unit tests