Remove RedactedReasoning from code (#566)
## Summary
Simplify the API by removing the need for "RedactedReasoning". Instead
we have a single Reasoning block, and if there's fields related to a
redaction, we keep them in the provider metadata.

This follows the recent changes in the Vercel SDK.

## How was it tested?
Updated tests and ran them.

## Community Contribution License

All community contributions in this pull request are licensed to the
project maintainers under the terms of the [Apache 2
License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the contributions to the project maintainers under the Apache 2
License as stated in the [Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).