Revert "[ai] add nil check for provider metadata (#568)" (#569)
## Summary

This reverts commit c6b33e126335c366c106eb5f28fcafab2887ae37.
That was a silly no-op

## How was it tested?

## Community Contribution License

All community contributions in this pull request are licensed to the
project maintainers under the terms of the [Apache 2
License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the contributions to the project maintainers under the Apache 2
License as stated in the [Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).