[aisdk] Allow passing of provider metadata in tool definitions (#579)
## Summary
We need it so that we can specify whether the tool schema validation
should be strict or not.

## How was it tested?
Used with testpilot and its custom tools. I'll send a PR on axiom to
show its usage too.

## Community Contribution License

All community contributions in this pull request are licensed to the
project maintainers under the terms of the [Apache 2
License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the contributions to the project maintainers under the Apache 2
License as stated in the [Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).